+++ app/data/__init__.py
@@ -14,7 +14,6 @@ DATA_SOURCES = {
 def data_source(source):
     """
     Retrieves the provided data-source service.
-
     :returns: The service.
     :rtype: LocationService
     """
